This is all it takes to make a new Endpoint in Jhipster! GOD!
Understanding the achitecture of Jhipster Back-End revealed to quite of
a task but it now resolved. I should be able to create my new endpoints
more easily now that I know! That did cost me a lot of motivation! but
whatever! the result is what counts Moffffoooos! hahaha it's beautiful
to succeed hard tasks! Up to the next one now!