Auto Id generation reactivated + new column
I realised that it was not a good idea to delete the auto id generation
for this entity since it does not have a one to one but one to many
relationship with the user entity. I hence reactivated the auto
generation and added a new column to contain the id of the user as a
foreign key.