Business logic added to save()
I did not want the user to be able to change its balance before valid
bank account information were added to his information.

This method is now checking => if the balance is greater than 0 => if
the balance changed from what was previously recorded => if yes, it will
then check if valid informations are in the bank account entity. If not,
it will throw an exception. If correct, it will proceed to the changes.